remove collocation checks for staging area links
After moving npg_tracking to containers in head nodes and once staging areas
are not local to head nodes, the check is not valid anymore. We expect staging
areas will be independent of farm head nodes (different filesystem) and we are
not allowing the container to access local filesystems where staging areas
would be.