Simplified run complete and mirrored detection.
Removed exessive complexity in finding RTACpmplete.txt
and CopyComplete.txt files.

Removed a second round of examining the above files
at the stage of deciding whether the mirroring has
been completed.

Removed examination of the content of the Error log
file since it is not present in the runfolders created
by the contemporary instruments.