Flagged planned onboard analysis.
The flag is implemented in the long_info role, which makes it
available to any class that inherits from this role, i.e. to
most of the code that deals with the run folders.

The implementation examines the content of RunParameters.xml

The new flag replaces the previous implementation of a similar
flag in the staging monitor.