Distinguish between NovaSeq and NovaSeqX RunParams.
We use the name of the Application in RunParameters.xml
to detect that the instrument belongs to the NovaSeq
type. Newer NovaSeqX instrument we have has the following
definition for the application:
  <Application>NovaSeqXSeries Control Software</Application>
This pull request fixes the reg exp we use to ensure that
the above definition does not result in the instrument being
classified as NovaSeq.