Ability to derive id_run from a run record
The old samplesheet generator had this ability.
Rearranged the code so that the samplesheet generator for
NovaSeqX also has this ability, which simplifies the code
in the samplesheet daemon. NovaSeqX samplesheet generator
retains ability to generate a samplesheet when neither
id_run nor the run db record is set via the constructor.