Stopped generating InstrumentFormat ORM class.
Dropped instrument_utilisation table from the db schema,
which is used for testing the tracking web application.

'instrument_utilisation' table has not been used for the
last seven years. The code that filled it in was deleted
some time ago.

Once this change is released, the 'instrument_utilisation'
table can be dropped from the prod. database.