Added sample_uuid, sample_lims methods to st::api::lims
Patched st::api::lims::samplesheet to allow for scalar attributes
ending with 's'. Previously the values of attributes with names
ending with 's' were automatically converted to lists. In fact,
this was necessary only for comma-separated strings of email addresses.
The part of code that deals with email addresses was changed to
work on attributes with names starting with 'email' rather than
ending with 's'. Prior to applying this patch the value of the
new sample_lims attribute was returned as a list when a samplesheet
was used as a source of LIMS data.

Updated tests and test data.