Refactored runs listing.
Added a new optional parameter to methods
returning lists of runs and run count - the name of
the manufacturer. If this name is undefined in the
parameters hash, it defaults to 'Illumina'.

To avoid code repetition, consolidated queries for
listing and counting runs, which previously were
in eight different places.

Simplified the logic of retrieving cached lists
and counts of runs vs computing new values.