Ensured consistency in handling paths
Ensured consistency in handling paths in
Monitor::Elembio::RunParametersParser

Changed the attribute type of both runparams_path and
runfolder_path so that the existence of the relevant file
or directory is performed.

Added builder methods for runparams_path and runfolder_path
attributes.

Ensured that runfolder_path attribute is used when a path of
the RunParameters.json file is needed.